um: add nommu mode for UML library mode
This patch introduces the nommu operation with UML code so that host
interface can be shrineked for broader environments support.

The nommu mode is implemneted as SUBARCH of arch/um, which places at
arch/um/nommu, which defines mode-specific code of memory managemnet as
well as thread implementation, along with the uapi headers to be
exported to users.

(TBA)

Signed-off-by: Hajime Tazaki <thehajime@gmail.com>
Signed-off-by: Octavian Purdila <tavi.purdila@gmail.com>