um: lkl: memory handling
library mode follows the way !CONFIG_MMU architecture of other linux
archs.  There is not much work left to do other than initializing the
boot allocator and providing the page and page table definitions.

The backstore memory is allocated via host functions and the memory
size to be used is specified when the kernel is started, in the
lkl_start_kernel call.

Signed-off-by: Hajime Tazaki <thehajime@gmail.com>
Signed-off-by: Octavian Purdila <tavi.purdila@gmail.com>