x86/um: nommu: add a workaround for MAP_FIXED
As !MMU mode MAP_FIXED flag is not supported, we add a workaround with
this commit.  The code is identical to the musl libc implementation,
which is enabled during compilation.

As we assume that we don't modify the original userspace binary, and
musl x86 binary does not include this work around (as x86 has no option
!MMU at this moment), we add this workaround at UML code.  But once the
userspace side (i.e., musl-libc) has a switch for the runtime
configuration (e.g., via env variable) and if we can turn on when we run
UML with !MMU, this patch can be eliminated in the future.

Signed-off-by: Hajime Tazaki <thehajime@gmail.com>